requirements.txt
requests         
openai-whisper  
python-dotenv
pydub
fastapi
uvicorn
python-multipart
yt_dlp

